+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 1.7.0
+
+Release date: work in progress 🚧
+
+* Playing a sound when rendering is finished or errors occur
+* Using the OS notification system when rendering is finished or errors occur
+
 ## Version 1.6.0
 
 Release date: 2021-01-09